requirements.txt
# Core dependencies
biopython>=1.79
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
pyyaml>=6.0
scikit-learn>=1.0.0

# Excel file handling
openpyxl>=3.0.0
bioservices>=1.11.0
requests>=2.28.0

# Optional: Jupyter notebooks
jupyter>=1.0.0
notebook>=6.4.0
matplotlib>=3.4.0
seaborn>=0.11.0
